feat(US-1.5): filter and sort delayed beds — all ACs complete
- Wire BedFilterBar into BedGridHeader: sort, active badge, clear
- BedGrid now passes all 7 useBedFilter values (was silently dropping 3)
- Add 14 US-1.5 AC integration tests (181/181 passing)

Architecture fix (ARCH-NOTE resolved):
- Move session.ts, auth-service.ts, active-session.ts to shared/lib/
- Replace feature originals with re-export stubs
- Update all 8 import sites to @/shared/lib/
- shared/ now has zero imports from features/

Bug fixes:
- migrations/014: comment out DROP TABLE in UP migration (kiosk_sessions was destroyed on create)
- run-migrations.js: move require to top (Cannot access before initialization)
- db-validators.ts: 'ward' -> 'wards' in ALLOWED_TABLES
- session.ts: 3 kiosk session bugs (renewSession drops fields, renewal resets expiry, inactivity on kiosk)

CONTRIBUTING.md compliance:
- kiosk-admin-actions.ts: remove cross-feature import, use requireAdmin + shared kiosk-queries
- realtime.ts: RealtimeConfig moved to shared/config
- Stray src/components/ and src/hooks/ removed, moved to src/shared/
- analytics-actions.ts (299L) split into 3 files
- stage-validation-rules.ts (208L) extracted consistency helper
- db-helpers.ts (206L) removed unused getById
- Dead code removed: ANALYTICS_EXAMPLE.tsx, getById, stale kiosk functions

Closes #10
