+++ src/app/supervisor/page.tsx
@@ -4,7 +4,7 @@ import { KioskBanner } from "@/features/auth/components/KioskBanner"
 import { redirect } from "next/navigation"
 import { AlertTriangle } from "lucide-react"
 
-import { verifyActiveSession } from "@/features/auth/lib/active-session"
+import { verifyActiveSession } from "@/shared/lib/active-session"
 import { getBedGridData } from "@/features/bed-dashboard/actions/bed-grid-actions"
 import { SupervisorBedOverview } from "@/features/bed-dashboard/components/SupervisorBedOverview"
