+++ src/features/user-management/actions/kiosk-admin-actions.ts
@@ -4,24 +4,25 @@
 // Epic 5: Authentication & Role-Based Access (US-5.3)
 // Purpose: Server actions for admin to list and revoke kiosk sessions
 
-import { verifyActiveSession } from '@/features/auth/lib/active-session'
-import { listActiveKioskSessions, revokeKioskSession } from '@/features/auth/lib/kiosk'
+import { requireAdmin } from '@/features/user-management/lib/auth'
+import { listActiveKioskSessions, revokeKioskSession } from '@/shared/lib/kiosk-queries'
+import { logger } from '@/shared/config/logger'
 
 /**
  * Returns all active kiosk sessions for the admin panel.
  * Admin-only.
  */
 export async function getActiveKioskSessionsAction() {
-  const session = await verifyActiveSession()
-  if (!session || session.role !== 'admin') {
-    return { success: false as const, error: 'Unauthorized', sessions: [] }
-  }
-
   try {
+    await requireAdmin()
     const sessions = await listActiveKioskSessions()
     return { success: true as const, sessions }
   } catch (error) {
-    console.error('Failed to list kiosk sessions:', error)
+    const message = error instanceof Error ? error.message : 'Failed to fetch kiosk sessions'
+    if (message.startsWith('Unauthorized')) {
+      return { success: false as const, error: 'Unauthorized', sessions: [] }
+    }
+    logger.error('Failed to list kiosk sessions', error as Error)
     return { success: false as const, error: 'Failed to fetch kiosk sessions', sessions: [] }
   }
 }
@@ -32,16 +33,16 @@ export async function getActiveKioskSessionsAction() {
  * Admin-only.
  */
 export async function revokeKioskSessionAction(kioskSessionId: string) {
-  const session = await verifyActiveSession()
-  if (!session || session.role !== 'admin') {
-    return { success: false as const, error: 'Unauthorized' }
-  }
-
   try {
+    const session = await requireAdmin()
     await revokeKioskSession(kioskSessionId, session.userId)
     return { success: true as const }
   } catch (error) {
-    console.error('Failed to revoke kiosk session:', error)
+    const message = error instanceof Error ? error.message : 'Failed to revoke session'
+    if (message.startsWith('Unauthorized')) {
+      return { success: false as const, error: 'Unauthorized' }
+    }
+    logger.error('Failed to revoke kiosk session', error as Error)
     return { success: false as const, error: 'Failed to revoke session' }
   }
 }
