+++ src/shared/lib/db-helpers.ts
@@ -40,32 +40,6 @@ export async function getAll<T = unknown>(
 }
 
 /**
- * Get a single record by ID
- * @param table - Table name
- * @param id - Record ID
- * @param idColumn - ID column name (default: 'id')
- */
-export async function getById<T = unknown>(
-  table: string,
-  id: string,
-  idColumn = 'id'
-): Promise<T | null> {
-  try {
-    validateTableName(table)
-    validateColumnName(idColumn)
-
-    const { rows } = await pool.query(
-      `SELECT * FROM ${table} WHERE ${idColumn} = $1`,
-      [id]
-    )
-    return rows.length > 0 ? (rows[0] as T) : null
-  } catch (error) {
-    logger.error(`Failed to get ${table} by ID`, error as Error)
-    throw error
-  }
-}
-
-/**
  * Check if a record exists
  * @param table - Table name
  * @param where - WHERE clause (without 'WHERE' keyword)
@@ -136,70 +110,6 @@ export async function reactivate(
     )
   } catch (error) {
     logger.error(`Failed to reactivate in ${table}`, error as Error)
-    throw error
-  }
-}
-
-/**
- * Generic update helper
- * @param table - Table name
- * @param id - Record ID
- * @param updates - Object containing fields to update
- * @param idColumn - ID column name (default: 'id')
- */
-export async function updateRecord(
-  table: string,
-  id: string,
-  updates: Record<string, unknown>,
-  idColumn = 'id'
-): Promise<void> {
-  try {
-    validateTableName(table)
-    validateColumnName(idColumn)
-
-    const fields = Object.keys(updates)
-    const values = Object.values(updates)
-    
-    // Validate all field names
-    fields.forEach(field => validateColumnName(field))
-
-    const setClause = fields
-      .map((field, index) => `${field} = $${index + 1}`)
-      .join(', ')
-
-    await pool.query(
-      `UPDATE ${table} SET ${setClause}, updated_at = NOW() WHERE ${idColumn} = $${fields.length + 1}`,
-      [...values, id]
-    )
-  } catch (error) {
-    logger.error(`Failed to update ${table}`, error as Error)
-    throw error
-  }
-}
-
-/**
- * Count records with optional filtering
- * @param table - Table name
- * @param where - Optional WHERE clause (without 'WHERE' keyword)
- * @param params - Query parameters
- */
-export async function count(
-  table: string,
-  where?: string,
-  params: unknown[] = []
-): Promise<number> {
-  try {
-    validateTableName(table)
-
-    let query = `SELECT COUNT(*) as count FROM ${table}`
-    if (where) {
-      query += ` WHERE ${where}`
-    }
-
-    const { rows } = await pool.query(query, params)
-    return parseInt(rows[0].count, 10)
-  } catch (error) {
-    logger.error(`Failed to count ${table}`, error as Error)
     throw error
   }
 }
